Fix the android example and update the documentation
